[수정] 클린아키텍쳐를 해당 프로젝트에 맞게끔 적용
