+++ UJeongApp/APP/Common/BaseArchitecture/BaseCollectionViewCell.swift
