+++ UJeongApp/APP/Scene/Notice/NoticeNavigator.swift
@@ -33,6 +33,3 @@ class DefaultNoticeNavigator: NoticeNavigator {
     func toNoticeDetail() { }
 }
 
-protocol UseCaseProvider {
-    func makeNoticeUseCase() -> NoticeUseCase
-}
