+++ UJeongApp/APP/SceneDelegate.swift
@@ -1,0 +1,23 @@
+//
+//  SceneDelegate.swift
+//  UJeongApp
+//
+//  Created by Hamlit Jason on 2022/12/30.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        
+        window = UIWindow(windowScene: windowScene)
+        window?.backgroundColor = .systemBackground
+        Application.shard.configureMainInterface(in: window)
+        window?.makeKeyAndVisible()
+    }
+}
