+++ UJeongApp/APP/SupportingFiles/Assets.xcassets/AccentColor.colorset/Contents.json
