+++ UJeongApp/Domain/Models/Notice.swift
