+++ UJeongApp/Domain/UseCases/UseCaseProvider.swift
@@ -1,0 +1,12 @@
+//
+//  UseCaseProvider.swift
+//  UJeongApp
+//
+//  Created by Hamlit Jason on 2023/01/01.
+//
+
+import Foundation
+
+protocol UseCaseProvider {
+    func makeNoticeUseCase() -> NoticeUseCase
+}
